+++ firmware/application/usb_serial_shell.hpp
